AI_WikiQuiz/backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
sqlalchemy
psycopg2-binary
langchain
python-dotenv
google-generativeai
